+++ CHANGELOG.md
@@ -19,6 +19,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Listen to keydown events on textarea for specific keys: 'tab', 'enter' and 'numpadenter'.
+
 ## [3.5.2](https://github.com/Smartesting/gravity-data-collector/compare/v3.5.1...v3.5.2)
 
 ### Fixed